Unify tensor dtypa to float32

+++ gc_bert/dataset.py
@@ -95,7 +95,7 @@ class PubmedDataset(Dataset):
 
         if to_sparse:
             self.adj = nx.convert_matrix.to_scipy_sparse_matrix(
-                self.G, nodelist=self.articles.index.tolist(), dtype='float64',
+                self.G, nodelist=self.articles.index.tolist(), dtype='float32',
             )
         else:
             self.adj = nx.convert_matrix.to_numpy_array(
